+++ app/controllers/parcours_controller.rb
@@ -5,7 +5,11 @@ class ParcoursController < ApplicationController
   # GET /parcours
   # GET /parcours.json
   def index
-    @parcours = Parcours.where(domaine_activite_parcours: params[:domaine])
+    if params[:acheteur]
+      @parcours = Parcours.par_acheteur(params[:acheteur])
+    else
+      @parcours = Parcours.where(domaine_activite_parcours: params[:domaine])
+    end
     @titre_parcours = params[:domaine]
   end
 
@@ -57,13 +61,13 @@ class ParcoursController < ApplicationController
   end
 
   private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_parcours
-      @parcours = Parcours.find(params[:id])
-    end
+  # Use callbacks to share common setup or constraints between actions.
+  def set_parcours
+    @parcours = Parcours.find(params[:id])
+  end
 
-    # Never trust parameters from the scary internet, only allow the white list through.
-    def parcours_params
-      params.require(:parcours).permit(:nom_organisme_acheteur, :type_acheteur, :adresse_acheteur, :nom_prenom_contact_acheteur, :mail_contact_acheteur, :telephone_contact_acheteur, :objet_marche, :date_notification_marche, :date_fin_clause_sociale, :date_fin_clause_sociale_reconduction_comprise, :nom_entreprise, :secteur_activite_entreprise, :site_web_entreprise, :nom_prenom_referent_entreprise, :telephone_referent_entreprise, :fonction_referent_entreprise, :nom_prenom_2eme_referent_entreprise, :telephone_2eme_referent_entreprise, :fonction_2eme_referent_entreprise, :heures_parcours, :adresse_parcours, :region_parcours, :departement_parcours, :academie_parcours, :domaine_activite_parcours, :details_taches_proposees_parcours, :autre_remarque)
-    end
+  # Never trust parameters from the scary internet, only allow the white list through.
+  def parcours_params
+    params.require(:parcours).permit(:nom_organisme_acheteur, :type_acheteur, :adresse_acheteur, :nom_prenom_contact_acheteur, :mail_contact_acheteur, :telephone_contact_acheteur, :objet_marche, :date_notification_marche, :date_fin_clause_sociale, :date_fin_clause_sociale_reconduction_comprise, :nom_entreprise, :secteur_activite_entreprise, :site_web_entreprise, :nom_prenom_referent_entreprise, :telephone_referent_entreprise, :fonction_referent_entreprise, :nom_prenom_2eme_referent_entreprise, :telephone_2eme_referent_entreprise, :fonction_2eme_referent_entreprise, :heures_parcours, :adresse_parcours, :region_parcours, :departement_parcours, :academie_parcours, :domaine_activite_parcours, :details_taches_proposees_parcours, :autre_remarque)
+  end
 end